+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 - Bugfix:           sp3Format2Orbit: no/invalid orbit positions/velocities are now excluded.
 - Bugfix:           Conversion of GRACE L1B/L1A data: revised source code.
 - Bugfix:           loopFileAscii: Fixed uninitialized variable that could lead to the loop ending prematurely.
+- Other:            File GriddedDataTimeSeries: includes now the last epoch; interval [...] instead of [...).
+- Other:            File TimeSplinesGravityfield: includes now the last epoch; interval [...] instead of [...).
 - Other:            Removed inputfileGlobal option.
 - Other:            GnssAttitude2Orbex: can now handle different sampling per satellite.
 - Other:            GnssRinexNavigation2OrbitClock/RinexObservation2GnssReceiver: Added basic support for RINEX v4.00.